+++ bridges/ironSorce/ironsorce-mediation-sample/app/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdkVersion 34
 
     defaultConfig {
-        applicationId "com.loopme.ironsource_76"
+        applicationId "com.loopme.ironsource_mediation_sample"
         minSdkVersion 24
         targetSdkVersion 34
         versionCode 1
@@ -25,7 +25,7 @@ android {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
-    namespace 'com.loopme.ironsource_76'
+    namespace 'com.loopme.ironsource_mediation_sample'
 }
 
 repositories {
