+++ bridges/ironSorce/ironsorce-mediation-sample/app/src/main/java/com/loopme/ironsource_mediation_sample/RewardedVideoActivity.java
@@ -1,4 +1,4 @@
-package com.loopme.ironsource_76;
+package com.loopme.ironsource_mediation_sample;
 
 import android.app.Activity;
 import android.os.Bundle;
